Update Cards.jsx
Imported meteor template, fixed lint issues in Cards.jx
